+++ examples/bookshelf/src/commonMain/kotlin/com/republicate/skorm/bookshelf/Aliases.kt
@@ -1,0 +1,11 @@
+package com.republicate.skorm.bookshelf
+
+// The generator only provides scoped names; an app aliases the ones it uses.
+
+typealias Author = ExampleDatabase.BookshelfSchema.Author
+typealias Book = ExampleDatabase.BookshelfSchema.Book
+typealias BookFields = ExampleDatabase.BookshelfSchema.BookFields
+typealias BookTag = ExampleDatabase.BookshelfSchema.BookTag
+typealias Dude = ExampleDatabase.BookshelfSchema.Dude
+typealias Genre = ExampleDatabase.BookshelfSchema.Genre
+typealias Tag = ExampleDatabase.BookshelfSchema.Tag